Remove useless variables from TileData, reduce mem usage of world.Object type...

+++ pkg/server/world/object.go
@@ -21,8 +21,8 @@ var Objects []ObjectDefinition
 
 //Object Represents a game object in the world.
 type Object struct {
-	ID        int
-	Direction int
+	ID        uint16
+	Direction byte
 	Boundary  bool
 	*Entity
 }
@@ -35,7 +35,7 @@ var ObjectCounter = atomic.NewUint32(0)
 
 //NewObject Returns a reference to a new instance of a game object.
 func NewObject(id, direction, x, y int, boundary bool) *Object {
-	return &Object{ID: id, Direction: direction, Boundary: boundary,
+	return &Object{ID: uint16(id), Direction: byte(direction), Boundary: boundary,
 		Entity: &Entity{
 			Location: NewLocation(x, y),
 			Index:    int(ObjectCounter.Swap(ObjectCounter.Load() + 1)),
